Function csc411_rpegio::output_rpeg_data

pub fn output_rpeg_data(raw_bytes: &[[u8; 4]], width: u32, height: u32)

Outputs rpeg data to stdout.

Arguments

raw_bytes - A slice of four-byte arrays, each array representing a single word of compressed image data

width - The width of the image

height - The height of the image

Panics

If something goes wrong writing raw bytes to stdout

Examples

// In your program, this rpeg data would be generated by compressing a .ppm file.
// Here, I've just made up some random data
let width: u32 = 2;
let height: u32 = 1;
let raw_bytes: Vec<[u8; 4]> = vec![[0x00, 0x11, 0x22, 0x33], [0x44, 0x55, 0x66, 0x77]];

// Output the rpeg data to stdout
csc411_rpegio::output_rpeg_data(&raw_bytes, width, height);