csc411_rpegio
Function output_rpeg_data_debug

pub fn output_rpeg_data_debug(
    compressed_data: &[[u8; 4]],
    width: usize,
    height: usize,
) -> Result<(), String>

Outputs rpeg data to stdout in a human-readable, debug form. This is only intended to be used for debugging and testing.

Arguments

compressed_data - A slice of four-byte arrays, where each array contains the big-endian bytes of a single word of compressed image data

width - The width of the image

height - The height of the image

Errors

If something goes wrong writing bytes to stdout

Examples

// In your program, this data would be generated by compressing a .ppm file.
// Here, I've just made up some random data
let width = 4;
let height = 2;
let compressed_data: Vec<[u8; 4]> = vec![[0x00, 0x11, 0x22, 0x33], [0x44, 0x55, 0x66, 0x77]];

// Output the rpeg data to stdout
csc411_rpegio::output_rpeg_data_debug(&compressed_data, width, height).unwrap();

// Standard Output:
// Compressed image format 2 [DEBUG]
// 4 2
// 00 11 22 33 44 55 66 77