Function csc411_rpegio::read_in_rpeg_data

pub fn read_in_rpeg_data(
    file_path: Option<&str>
) -> Result<(Vec<[u8; 4]>, u32, u32), String>

Reads and parses rpeg data from either stdin or a file. Returns a Result<tuple, String> where the tuple contains, in order:

A Vec<[u8; 4]> (Vector of four-byte arrays) representing the raw image data

A u32 representing the width of the image

A u32 representing the height of the image

Errors Returned

If there is an unexpected error reading from the provided file or stdin

If the rpeg data header is badly formatted

If the number of raw bytes following the header is not a multiple of 4

Arguments

file_path - An optional file path to read from. If None, stdin will be read from instead

Examples

// Read rpeg data from stdin to variables for later use
let (raw_bytes, width, height) = csc411_rpegio::read_in_rpeg_data(Some("path/to/file.ppm")).unwrap();

// Do something with width and height. This is just an example
println!("Image size: {width}x{height}");

// Do something with raw_bytes
// you will likely first want to convert the four-byte arrays to u32s