Crate csc411_rpegio

csc411_rpegio

A collection functions to handle rpeg data i/o. Intended for use in URI’s CSC 411 class.

Functions

input_rpeg_data

Reads and parses rpeg data from a file, or stdin if no file path is given.

output_rpeg_data

Outputs rpeg data to stdout.

output_rpeg_data_debug

Outputs rpeg data to stdout in a human-readable, debug form. This is only intended to be used for debugging and testing.