Function csc411_rpegio::input_rpeg_data

pub fn input_rpeg_data<P: AsRef<Path>>(
    file_path: Option<P>
) -> Result<(Vec<[u8; 4]>, usize, usize), String>

Reads and parses rpeg data from a file, or stdin if no file path is given.

Returns a Result<tuple, String> where the tuple contains, in order:

A Vec<[u8; 4]> containing the compressed data bytes of the image

A usize for the width of the image

A usize for the height of the image

Errors

If there is an error reading from the provided file or stdin

If the rpeg data is invalid (ex: malformed header, mismatched size, etc.)

Examples

// Read in the rpeg data
let input = Some("path/to/file.rpeg");
let (compressed_data, width, height) = csc411_rpegio::input_rpeg_data(input).unwrap();

// Do something with width and height. This example just prints them
println!("Image size: {width}x{height}");

// Do something with `compressed_data`
// You will likely want to interpret the `[u8; 4]`'s as big-endian `u32`s
let words: Vec<u32> = compressed_data.into_iter().map(u32::from_be_bytes).collect();